Repo Setup (#1)
* Create README.md

* chore: gitignore for swift

* chore: add .gitattributes

The line in .gitattributes treats your Xcode project file as a binary. This prevents Git from trying to fix newlines, show it in diffs, and excludes it from merges. Note that you will still see it shown as a conflict in merges, although the file won’t have changed. Simply commit it and things should be good.